FlaskApp/requirements.txt
Flask
flask-login
pandas
dash
dash-bootstrap-components
Flask-Caching
oauthlib
dropbox
ftputil
tabulate
pyopenssl
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
flask-sqlalchemy
flask-migrate
openpyxl